test: exercise sticker FIFO rejection through the CLI
