+++ Tests/IMsgCoreTests/StickerAssetTests.swift
@@ -1,5 +1,4 @@
 import CoreGraphics
-import Darwin
 import Foundation
 import ImageIO
 import Testing
@@ -111,32 +110,6 @@ func stickerPreparationRejectsUnsafeOrInvalidInputs() throws {
   #expect(throws: StickerAssetError.self) {
     try StickerAssetPreparer.prepare(at: root.path, destinationRoot: staged)
   }
-}
-
-@Test
-func stickerPreparationRejectsFIFOWithoutWaitingForAWriter() throws {
-  let root = FileManager.default.temporaryDirectory
-    .appendingPathComponent(UUID().uuidString, isDirectory: true)
-  try FileManager.default.createDirectory(at: root, withIntermediateDirectories: true)
-  defer { try? FileManager.default.removeItem(at: root) }
-  let fifo = root.appendingPathComponent("sticker.png")
-  #expect(mkfifo(fifo.path, 0o600) == 0)
-
-  let completed = DispatchSemaphore(value: 0)
-  DispatchQueue.global().async {
-    defer { completed.signal() }
-    #expect(throws: StickerAssetError.self) {
-      try StickerAssetPreparer.prepare(at: fifo.path)
-    }
-  }
-  let rejectedWithoutWriter = completed.wait(timeout: .now() + 2) == .success
-  // Release a regressed blocking open before removing the fixture or ending the test.
-  if !rejectedWithoutWriter {
-    let writer = open(fifo.path, O_WRONLY | O_NONBLOCK)
-    if writer >= 0 { close(writer) }
-    #expect(completed.wait(timeout: .now() + 2) == .success)
-  }
-  #expect(rejectedWithoutWriter)
 }
 
 @Test
